Run buffer cleanup test in Emacs >= 30 only.
I have no reasons to believe this code does not run properly
in older than version 30 Emacsen, but because they reliably
trigger `image-mode`, the test spits the dummy because it
is run in batch mode.

We should safely assume that outside of that detail, the code
should work in these versions, and pass the CI tests.